requirements.txt
django==2.0
mysqlclient==1.3.12
djangorestframework==3.7.3